[tests] Move unique_id_mapping check to new unit.
